ref(rcs): split presence handling between tv and remote
